親子関係保持処理をInventoriableObjectに移動
