+++ Assets/SSP/Prefabs/GameManager.prefab
@@ -19,9 +19,11 @@ GameObject:
   serializedVersion: 5
   m_Component:
   - component: {fileID: 4827270185244938}
+  - component: {fileID: 114257975132688072}
   - component: {fileID: 114995220335104442}
   - component: {fileID: 114974705603106214}
   - component: {fileID: 114968024861136860}
+  - component: {fileID: 114250023417241802}
   m_Layer: 0
   m_Name: GameManager
   m_TagString: Untagged
@@ -42,6 +44,50 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &114250023417241802
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1452514462489488}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 37c2b38096c992947aeed1e2a9179375, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  text: {fileID: 0}
+--- !u!114 &114257975132688072
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1452514462489488}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 372142912, guid: dc443db3e92b4983b9738c1131f555cb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_SceneId:
+    m_Value: 0
+  m_AssetId:
+    i0: 185
+    i1: 221
+    i2: 97
+    i3: 203
+    i4: 150
+    i5: 25
+    i6: 131
+    i7: 100
+    i8: 233
+    i9: 71
+    i10: 58
+    i11: 188
+    i12: 227
+    i13: 250
+    i14: 210
+    i15: 235
+  m_ServerOnly: 1
+  m_LocalPlayerAuthority: 0
 --- !u!114 &114968024861136860
 MonoBehaviour:
   m_ObjectHideFlags: 1
