+++ Assets/SSP/Prefabs/Weapons/Bat.prefab
@@ -25,7 +25,6 @@ GameObject:
   - component: {fileID: 114506838946006212}
   - component: {fileID: 114244314045166808}
   - component: {fileID: 114645791642823236}
-  - component: {fileID: 114931751636707102}
   m_Layer: 0
   m_Name: Bat
   m_TagString: Untagged
@@ -144,19 +143,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   inventoriableType: 0
   canInteract: 1
---- !u!114 &114931751636707102
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1189635630215696}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 297c3b8214708f445914ab66d8cdf4ad, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  playerId: 0
-  teamId: 0
 --- !u!136 &136532114692406812
 CapsuleCollider:
   m_ObjectHideFlags: 1
