+++ Assets/SSP/Prefabs/Weapons/GimmickGun.prefab
@@ -73,7 +73,6 @@ GameObject:
   - component: {fileID: 114729662647059564}
   - component: {fileID: 114128460819389056}
   - component: {fileID: 114250481760157286}
-  - component: {fileID: 114966800734412244}
   m_Layer: 0
   m_Name: GimmickGun
   m_TagString: Untagged
@@ -386,16 +385,3 @@ MonoBehaviour:
     i15: 130
   m_ServerOnly: 0
   m_LocalPlayerAuthority: 1
---- !u!114 &114966800734412244
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1595191971914448}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 297c3b8214708f445914ab66d8cdf4ad, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  playerId: 0
-  teamId: 0
