+++ Assets/SSP/Prefabs/Weapons/HandGun.prefab
@@ -58,7 +58,6 @@ GameObject:
   - component: {fileID: 114397692148794256}
   - component: {fileID: 114275136129540490}
   - component: {fileID: 114611022078819548}
-  - component: {fileID: 114604629963241754}
   m_Layer: 0
   m_Name: HandGun
   m_TagString: Untagged
@@ -261,19 +260,6 @@ MonoBehaviour:
     i15: 25
   m_ServerOnly: 0
   m_LocalPlayerAuthority: 1
---- !u!114 &114604629963241754
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1996856847404154}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 297c3b8214708f445914ab66d8cdf4ad, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  playerId: 0
-  teamId: 0
 --- !u!114 &114611022078819548
 MonoBehaviour:
   m_ObjectHideFlags: 1
