--- Assets/SSP/Scripts/Weapons/WeaponModel.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 297c3b8214708f445914ab66d8cdf4ad
-timeCreated: 1507271601
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
