+++ Assets/SSP/Scripts/items/InventoriableObject.cs
@@ -1,12 +1,21 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Networking;
 
 public enum InventoriableType { HandGun, LongRangeWeapon, ShortRangeWeapon, Gimmick }
-public class InventoriableObject : MonoBehaviour, IInteractable
+public class InventoriableObject : NetworkBehaviour, IInteractable
 {
     [SerializeField] private InventoriableType inventoriableType;
     [SerializeField] private bool canInteract = true;
+    
+    [SyncVar] public NetworkInstanceId ownerPlayerId;
+
+    void Start()
+    {
+        if (ownerPlayerId != null)
+            SetParent();
+    }
 
     public void Interact(PlayerManager pm)
     {
@@ -22,5 +31,14 @@ public class InventoriableObject : MonoBehaviour, IInteractable
     public void SetCanInteract(bool _canInteract)
     {
         canInteract = _canInteract;
+    }
+
+    [ClientCallback]
+    private void SetParent()
+    {
+        var player = ClientScene.FindLocalObject(ownerPlayerId);
+        if (player == null) return;
+        transform.SetParent(player.GetComponent<PlayerInventoryManager>().rightHandTransform);
+        transform.localPosition = Vector3.zero;
     }
 }
